chore: move wide-logo out of shadcn folder
