+++ app/components/wide-logo.tsx
